refactor(filecoin-proofs): split out post api
