+++ filecoin-proofs/src/api/mod.rs
@@ -1,4 +1,3 @@
-use std::collections::HashMap;
 use std::convert::TryInto;
 use std::fs::{copy, File, OpenOptions};
 use std::io::prelude::*;
@@ -8,11 +7,8 @@ use std::path::{Path, PathBuf};
 use memmap::MmapOptions;
 use paired::bls12_381::Bls12;
 use paired::Engine;
-use rayon::prelude::*;
 
-use crate::caches::{
-    get_post_params, get_post_verifying_key, get_zigzag_params, get_zigzag_verifying_key,
-};
+use crate::caches::{get_zigzag_params, get_zigzag_verifying_key};
 use crate::constants::{
     MINIMUM_RESERVED_BYTES_FOR_PIECE_IN_FULLY_ALIGNED_SECTOR as MINIMUM_PIECE_SIZE,
     MINIMUM_RESERVED_LEAVES_FOR_PIECE_IN_SECTOR as MIN_NUM_LEAVES, POREP_MINIMUM_CHALLENGES,
@@ -21,20 +17,18 @@ use crate::constants::{
 use crate::error;
 use crate::file_cleanup::FileCleanup;
 use crate::fr32::{write_padded, write_unpadded};
-use crate::parameters::{post_setup_params, public_params, setup_params};
+use crate::parameters::{public_params, setup_params};
 use crate::pieces::{get_aligned_source, get_piece_alignment, PieceAlignment};
 use crate::singletons::ENGINE_PARAMS;
 use crate::types::{
-    PaddedBytesAmount, PoRepConfig, PoRepProofPartitions, PoStConfig, SectorSize,
-    UnpaddedByteIndex, UnpaddedBytesAmount,
+    PaddedBytesAmount, PoRepConfig, PoRepProofPartitions, SectorSize, UnpaddedByteIndex,
+    UnpaddedBytesAmount,
 };
 
 use storage_proofs::circuit::multi_proof::MultiProof;
-use storage_proofs::circuit::rational_post::RationalPoStCompound;
 use storage_proofs::circuit::zigzag::ZigZagCompound;
 use storage_proofs::compound_proof::{self, CompoundProof};
-use storage_proofs::drgraph::{DefaultTreeHasher, Graph};
-use storage_proofs::error::Error;
+use storage_proofs::drgraph::DefaultTreeHasher;
 use storage_proofs::fr32::{bytes_into_fr, fr_into_bytes, Fr32Ary};
 use storage_proofs::hasher::pedersen::{PedersenDomain, PedersenHasher};
 use storage_proofs::hasher::{Domain, Hasher};
@@ -45,11 +39,12 @@ use storage_proofs::piece_inclusion_proof::{
     PieceSpec,
 };
 use storage_proofs::porep::{replica_id, PoRep, Tau};
-use storage_proofs::proof::NoRequirements;
-use storage_proofs::rational_post;
-use storage_proofs::sector::*;
+use storage_proofs::sector::SectorId;
 use storage_proofs::zigzag_drgporep::ZigZagDrgPoRep;
 use tempfile::tempfile;
+
+mod post;
+pub use crate::api::post::*;
 
 /// FrSafe is an array of the largest whole number of bytes guaranteed not to overflow the field.
 pub type FrSafe = [u8; 31];
@@ -68,54 +63,6 @@ pub struct SealOutput {
     pub piece_inclusion_proofs: Vec<PieceInclusionProof<PedersenHasher>>,
 }
 
-/// The minimal information required about a replica, in order to be able to generate
-/// a PoSt over it.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct ReplicaInfo {
-    /// The id of this sector.
-    sector_id: SectorId,
-    /// Path to the replica.
-    access: String,
-    /// The replica commitment.
-    commitment: Commitment,
-}
-
-impl std::cmp::Ord for ReplicaInfo {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.commitment.as_ref().cmp(other.commitment.as_ref())
-    }
-}
-
-impl std::cmp::PartialOrd for ReplicaInfo {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl ReplicaInfo {
-    pub fn new(id: SectorId, access: String, commitment: Commitment) -> Self {
-        ReplicaInfo {
-            sector_id: id,
-            access,
-            commitment,
-        }
-    }
-
-    pub fn safe_commitment(&self) -> Result<PedersenDomain, failure::Error> {
-        as_safe_commitment(&self.commitment, "comm_r")
-    }
-
-    /// Generate the merkle tree of this particular replica.
-    pub fn merkle_tree(&self, sector_size: u64) -> Result<Tree, Error> {
-        let mut f_in = File::open(&self.access)?;
-        let mut data = Vec::new();
-        f_in.read_to_end(&mut data)?;
-
-        let bytes = PaddedBytesAmount(sector_size as u64);
-        public_params(bytes, 1).graph.merkle_tree(&data)
-    }
-}
-
 fn as_safe_commitment(
     comm: &Commitment,
     commitment_name: impl AsRef<str>,
@@ -127,179 +74,6 @@ fn as_safe_commitment(
             err
         )
     })
-}
-
-/// Generates a proof-of-spacetime.
-/// Accepts as input a challenge seed, configuration struct, and a vector of
-/// sealed sector file-path plus CommR tuples, as well as a list of faults.
-pub fn generate_post(
-    post_config: PoStConfig,
-    challenge_seed: &ChallengeSeed,
-    replicas: &[ReplicaInfo],
-    sectors: &SectorSet,
-    faults: &SectorSet,
-) -> error::Result<Vec<u8>> {
-    let sector_count = replicas.len() as u64;
-    let sector_size = u64::from(PaddedBytesAmount::from(post_config));
-
-    // Only one fault per sector is possible.
-    ensure!(
-        faults.len() as u64 <= sector_count,
-        "Too many faults submitted"
-    );
-
-    let vanilla_params = post_setup_params(post_config);
-    let setup_params = compound_proof::SetupParams {
-        vanilla_params: &vanilla_params,
-        engine_params: &(*ENGINE_PARAMS),
-        partitions: None,
-    };
-
-    let pub_params: compound_proof::PublicParams<_, rational_post::RationalPoSt<PedersenHasher>> =
-        RationalPoStCompound::setup(&setup_params)?;
-
-    let challenges = rational_post::derive_challenges(
-        vanilla_params.challenges_count,
-        sector_size,
-        sectors,
-        challenge_seed,
-        faults,
-    );
-
-    // Match the replicas to the challenges, as these are the only ones required.
-    let challenged_replicas: Vec<_> = challenges
-        .iter()
-        .map(|c| {
-            if let Some(replica) = replicas.iter().find(|r| r.sector_id == c.sector) {
-                Ok(replica)
-            } else {
-                Err(format_err!(
-                    "Invalid challenge generated: {}, only {} sectors are being proven",
-                    c.sector,
-                    sector_count
-                ))
-            }
-        })
-        .collect::<Result<_, _>>()?;
-
-    // Generate merkle trees for the challenged replicas.
-    // Merkle trees should be generated only once, not multiple times if the same sector is challenged
-    // multiple times, so we build a HashMap of trees.
-
-    let mut unique_challenged_replicas = challenged_replicas.clone();
-    unique_challenged_replicas.sort_unstable(); // dedup requires a sorted list
-    unique_challenged_replicas.dedup();
-
-    let unique_trees_res: Vec<_> = unique_challenged_replicas
-        .into_par_iter()
-        .map(|replica| replica.merkle_tree(sector_size).map(|tree| (replica, tree)))
-        .collect();
-
-    // resolve results
-    let unique_trees: HashMap<&ReplicaInfo, Tree> =
-        unique_trees_res.into_iter().collect::<Result<_, _>>()?;
-
-    let borrowed_trees: Vec<&Tree> = challenged_replicas
-        .iter()
-        .map(|replica| {
-            // Safe to unwrap, as we constructed the HashMap such that each of these has an entry.
-            unique_trees.get(*replica).unwrap()
-        })
-        .collect();
-
-    // Construct the list of actual commitments
-    let commitments: Vec<_> = challenged_replicas
-        .iter()
-        .map(|replica| replica.safe_commitment())
-        .collect::<Result<_, _>>()?;
-
-    let pub_inputs = rational_post::PublicInputs {
-        challenges: &challenges,
-        commitments: &commitments,
-        faults,
-    };
-
-    let priv_inputs = rational_post::PrivateInputs::<PedersenHasher> {
-        trees: &borrowed_trees[..],
-    };
-
-    let groth_params = get_post_params(post_config)?;
-
-    let proof = RationalPoStCompound::prove(&pub_params, &pub_inputs, &priv_inputs, &groth_params)?;
-
-    Ok(proof.to_vec())
-}
-
-/// Verifies a proof-of-spacetime.
-pub fn verify_post(
-    post_config: PoStConfig,
-    comm_rs: Vec<Commitment>,
-    challenge_seed: &ChallengeSeed,
-    proof: &[u8],
-    sectors: &SectorSet,
-    faults: &SectorSet,
-) -> error::Result<bool> {
-    let sector_size = u64::from(PaddedBytesAmount::from(post_config));
-    let sector_count = comm_rs.len() as u64;
-
-    // Only one fault per sector is possible.
-    ensure!(faults.len() as u64 <= sector_count, "Too many faults");
-
-    let vanilla_params = post_setup_params(post_config);
-    let setup_params = compound_proof::SetupParams {
-        vanilla_params: &vanilla_params,
-        engine_params: &(*ENGINE_PARAMS),
-        partitions: None,
-    };
-    let commitments_all: Vec<PedersenDomain> = comm_rs
-        .iter()
-        .map(|c| as_safe_commitment(c, "comm_r"))
-        .collect::<Result<_, _>>()?;
-
-    let challenges = rational_post::derive_challenges(
-        vanilla_params.challenges_count,
-        sector_size,
-        sectors,
-        challenge_seed,
-        faults,
-    );
-
-    // Match the replicas to the challenges, as these are the only ones required.
-    let commitments: Vec<_> = challenges
-        .iter()
-        .map(|c| {
-            if let Some(comm) = commitments_all.get(u64::from(c.sector) as usize) {
-                Ok(*comm)
-            } else {
-                Err(format_err!(
-                    "Invalid challenge generated: {}, only {} sectors are being proven",
-                    c.sector,
-                    sector_count
-                ))
-            }
-        })
-        .collect::<Result<_, _>>()?;
-
-    let public_params: compound_proof::PublicParams<
-        _,
-        rational_post::RationalPoSt<PedersenHasher>,
-    > = RationalPoStCompound::setup(&setup_params)?;
-
-    let public_inputs = rational_post::PublicInputs::<PedersenDomain> {
-        challenges: &challenges,
-        commitments: &commitments,
-        faults: &faults,
-    };
-
-    let verifying_key = get_post_verifying_key(post_config)?;
-
-    let proof = MultiProof::new_from_reader(None, &proof[..], &verifying_key)?;
-
-    let is_valid =
-        RationalPoStCompound::verify(&public_params, &public_inputs, &proof, &NoRequirements)?;
-
-    // Since callers may rely on previous mocked success, just pretend verification succeeded, for now.
-    Ok(is_valid)
 }
 
 struct PseudoPieceSpec {
@@ -682,11 +456,12 @@ fn pad_safe_fr(unpadded: &FrSafe) -> Fr32Ary {
 mod tests {
     use crate::constants::TEST_SECTOR_SIZE;
     use crate::error::ExpectWithBacktrace;
-    use crate::types::SectorSize;
+    use crate::types::{PoStConfig, SectorSize};
 
     use rand::Rng;
     use tempfile::NamedTempFile;
 
+    use storage_proofs::sector::SectorSet;
     use storage_proofs::util::NODE_SIZE;
 
     use super::*;
